Original change: https://android-review.googlesource.com/c/platform/external/perfetto/+/1594259

MUST ONLY BE SUBMITTED BY AUTOMERGER

+++ src/trace_processor/importers/proto/proto_trace_parser.cc
@@ -477,7 +477,7 @@ void ProtoTraceParser::ParsePerfSample(
   // requiring stack samples, then this would still count as an error.
   // For now, use an invalid callsite id.
   if (!cs_id) {
-    cs_id = base::make_optional<CallsiteId>(-1u);
+    cs_id = base::make_optional<CallsiteId>(static_cast<uint32_t>(-1));
   }
 
   UniqueTid utid =
